Improve ReadMes and Code Flow with latest refactor of scenarios
Note: generate_dataset.py is still being converted to the latest setup.

Change generate_dataset to comply with the changes

Address yapf formatting

Delete cpp-package.yml

Fix yapf formatting

Update markdown and gifs

Update fluid_block.gif to higher resolution

Improve Dataset Readme
