+++ generate_dataset.py
@@ -1,7 +1,7 @@
 """Generate a dataset of FluidBlock simulations.
 
 The dataset contains variations of the viscosity, position, dimensions
-and initial velocity of the block of fluid as it is released in the tank.
+and initial velocity of the block of fluid as it is released into the tank.
 
 The dataset is generated by launching multiple simulations in parallel
 using a MachineGroup. The MachineGroup is created with spot instances
@@ -13,7 +13,8 @@ import os
 import numpy as np
 
 import inductiva
-import fluid_block
+from lib import models
+from lib import scenarios
 
 # Constant parameters for all simulations
 DENSITY = 1e3  # in kg/m^3
@@ -21,7 +22,7 @@ TANK_DIMENSIONS = [1., 1., 1.]  # in meters
 SIMULATION_TIME = 3.  # in seconds
 PARTICLE_RADIUS = 0.008  # in meters
 OUTPUT_DIR = "dataset"
-N_SIMULATIONS = 100
+N_SIMULATIONS = 10
 
 # Generate dataset with FluidBlock scenario
 dataset_tasks = []
@@ -30,7 +31,11 @@ dataset_machines = inductiva.resources.MachineGroup(
     machine_type="c2-standard-4", num_machines=10, spot=True)
 dataset_machines.start()
 
+simulation_parameters = scenarios.SimulationParameters(
+    simulation_time=SIMULATION_TIME, particle_radius=PARTICLE_RADIUS)
+
 for i in range(N_SIMULATIONS):
+
     position = np.random.uniform(low=[0., 0., 0.],
                                  high=[0.5, 0.5, 0.5],
                                  size=(3,))
@@ -42,16 +47,13 @@ for i in range(N_SIMULATIONS):
     velocity = np.random.uniform(low=-1., high=1., size=(3,))
     viscosity = 10**np.random.uniform(low=-6., high=1.)
 
-    print(f"\n========== LAUNCHING SIMULATION {i}/{N_SIMULATIONS} ==========")
-    scenario = fluid_block.FluidBlock(density=DENSITY,
-                                      kinematic_viscosity=viscosity,
-                                      dimensions=dimensions,
-                                      position=position,
-                                      initial_velocity=velocity)
+    # Initialize Block parameters
+    block = models.FluidBlock(DENSITY, viscosity, dimensions, position,
+                              velocity)
 
-    task = scenario.simulate(simulation_time=SIMULATION_TIME,
-                             particle_radius=PARTICLE_RADIUS,
-                             particle_sorting=False,
+    print(f"\n========== LAUNCHING SIMULATION {i}/{N_SIMULATIONS} ==========")
+    scenario = scenarios.FluidBlockSplishSplash(block)
+    task = scenario.simulate(sim_params=simulation_parameters,
                              machine_group=dataset_machines)
 
     dataset_tasks.append(task)
