+++ lib/templates/splishsplash/unit_box.obj
@@ -1,0 +1,34 @@
+v 0.000000 0.000000 1.000000
+v 1.000000 0.000000 1.000000
+v 0.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 0.000000 1.000000 0.000000
+v 1.000000 1.000000 0.000000
+v 0.000000 0.000000 0.000000
+v 1.000000 0.000000 0.000000
+vt 0.375000 0.000000
+vt 0.625000 0.000000
+vt 0.375000 0.250000
+vt 0.625000 0.250000
+vt 0.375000 0.500000
+vt 0.625000 0.500000
+vt 0.375000 0.750000
+vt 0.625000 0.750000
+vt 0.375000 1.000000
+vt 0.625000 1.000000
+vt 0.875000 0.000000
+vt 0.875000 0.250000
+vt 0.125000 0.000000
+vt 0.125000 0.250000
+f 1/1 2/2 3/3
+f 3/3 2/2 4/4
+f 3/3 4/4 5/5
+f 5/5 4/4 6/6
+f 5/5 6/6 7/7
+f 7/7 6/6 8/8
+f 7/7 8/8 1/9
+f 1/9 8/8 2/10
+f 2/2 8/11 4/4
+f 4/4 8/11 6/12
+f 7/13 1/1 5/14
+f 5/14 1/1 3/3
