--- sph/__init__.py
@@ -1,5 +1,0 @@
-#pylint: disable=missing-module-docstring
-import logging
-
-logger = logging.getLogger()
-logger.setLevel(logging.INFO)
